requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
xgboost>=1.5.0
lightgbm>=3.3.0
elevation>=1.1.0
rasterio>=1.2.0
geopandas>=0.10.0
shapely>=1.8.0
requests>=2.26.0
flask>=2.0.0
geopy>=2.2.0
folium>=0.12.0
schedule>=1.1.0
pyyaml>=5.4.0
matplotlib>=3.4.0
seaborn>=0.11.0
joblib>=1.1.0
scipy>=1.7.0
